requirements.txt
kafka-python
pyspark
pymongo
kafka-python
pyspark
python-dotenv